create session config class for sharing enviroment config
