--- src/models/sessionStoreConfig.ts
@@ -1,6 +1,0 @@
-export interface SessionStoreConfig {
-    sessionCookieName: string;
-    sessionCookieDomain: string;
-    sessionCookieSecureFlag: string;
-    sessionTimeToLiveInSeconds: number;
-}
